Move process lifecycle helpers into core
Constraint: Idle timeout and sync shutdown govern OpenChrome host process lifecycle and should live with parent liveness handling.\nRejected: Keeping them under src/utils | they are runtime lifecycle policy, not generic helper code.\nRejected: Moving sync-shutdown under src/chrome | it is triggered by host process exits even though it calls Chrome-specific cleanup.\nConfidence: high\nScope-risk: moderate\nDirective: Keep host process lifecycle primitives in src/core/process; keep Chrome PID kill primitives in src/chrome.\nTested: npx jest --config jest.ci.config.js --runInBand --silent --runTestsByPath tests/core/process/idle-timeout.test.ts tests/core/process/sync-shutdown.test.ts tests/integration/embedded-server.test.ts tests/integration/idle-timeout.test.ts tests/integration/shutdown-reentrancy.test.ts\nTested: npm run build\nTested: npm run lint\nTested: npm run lint:tier\nTested: npm run lint:repo-structure\nTested: git diff --check\nNot-tested: Full Jest suite.
