+++ src/index.ts
@@ -20,7 +20,7 @@ import { resolveWindowBoundsConfig } from './config/window-bounds';
 import { ToolTier } from './config/tool-tiers';
 import { writePidFile, cleanOrphanedChromeProcesses } from './chrome/pid-manager';
 import { installParentWatcher, ParentWatcherHandle } from './core/process/parent-watcher';
-import { installIdleTimeout, IdleTimeoutHandle, parseDuration } from './utils/idle-timeout';
+import { installIdleTimeout, IdleTimeoutHandle, parseDuration } from './core/process/idle-timeout';
 import { getIdleState } from './utils/idle-state';
 import { getVersion } from './version';
 import { bootstrapPilot, logActiveFlags } from './harness/flags';
